certificate file should be encrypt by RSA
