+++ 2 Security/backend/user-manager/src/main/java/com/bitforcestudio/usermanager/configuration/ServerConfig.java
@@ -1,0 +1,51 @@
+package com.bitforcestudio.usermanager.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import org.apache.catalina.Context;
+import org.apache.catalina.connector.Connector;
+import org.apache.tomcat.util.descriptor.web.SecurityCollection;
+import org.apache.tomcat.util.descriptor.web.SecurityConstraint;
+
+import org.springframework.boot.web.embedded.tomcat.TomcatServletWebServerFactory;
+import org.springframework.boot.web.servlet.server.ServletWebServerFactory;
+
+@Configuration
+public class ServerConfig {
+
+    @Bean
+    public ServletWebServerFactory servletContainer() {
+        // Enable SSL Trafic
+        TomcatServletWebServerFactory tomcat = new TomcatServletWebServerFactory() {
+            @Override
+            protected void postProcessContext(Context context) {
+                SecurityConstraint securityConstraint = new SecurityConstraint();
+                securityConstraint.setUserConstraint("CONFIDENTIAL");
+                SecurityCollection collection = new SecurityCollection();
+                collection.addPattern("/user/*");
+                securityConstraint.addCollection(collection);
+                context.addConstraint(securityConstraint);
+            }
+        };
+
+        // Add HTTP to HTTPS redirect
+        tomcat.addAdditionalTomcatConnectors(httpToHttpsRedirectConnector());
+
+        return tomcat;
+    }
+
+    /*
+    We need to redirect from HTTP to HTTPS. Without SSL, this application used
+    port 8082. With SSL it will use port 8443. So, any request for 8002 needs to be
+    redirected to HTTPS on 8443.
+     */
+    private Connector httpToHttpsRedirectConnector() {
+        Connector connector = new Connector(TomcatServletWebServerFactory.DEFAULT_PROTOCOL);
+        connector.setScheme("http");
+        connector.setPort(8002);
+        connector.setSecure(false);
+        connector.setRedirectPort(8443);
+        return connector;
+    }
+}
